requirements.txt
openai-whisper==20231117
pip==24.0
pydub==0.25.1
PyQt6==6.7.0
spleeter==2.4.0
syncedlyrics==0.10.1
